+++ CHANGELOG.md
@@ -6,6 +6,7 @@ principles.
 
 ## unreleased
 - Fixed: Trigger PageChanged event when fragments are sorted.
+- Fixed: stay in sidebar if redirect is set for a fragment. By default a fragment with nested fragments redirects after creation to its edit page.
 
 ## 0.7.6 - 2022-05-11
 - Added: Simple state to use when your model only needs: online, offline and delete states. Add the `UsesSimpleState` trait.